refactor(exchange): extract executeSwap from the swap rpc handler
